Make RabbitServer's Queues and Exchanges public.
Consequently make other types public that would otherwise result in compile errors due to inconsistent visibilities.

The motivation for this is that without exposing the server's Queues and Exchanges, it's impossible (or at least impractical to the point where I don't understand how) to assert about state in unit tests, which renders the library unusable for its intended purpose.